Address PR review: move format_personality out of node base; reword prompt (NPC-683)
- format_personality moved from nodes/base.py to a new nodes/formatting.py.
  base.py is the generic LLMNode base; personality prompt-rendering is
  domain logic and does not belong there (per review). Both consumers
  (cognitive_update, action_selection) updated to import from the new
  module. Behavior unchanged — pure relocation.
- cognitive_update/prompt.md: "situation read" -> "situation assessment"
  to match the WorkingMemory.situation_assessment field name (gemini nit).
