+++ mind/src/mind/cognitive_architecture/nodes/cognitive_update/prompt.md
@@ -36,7 +36,7 @@ Update the working memory with:
 4. **Current plan** - Their intended next steps to achieve goals
 5. **Emotional state** - How they're feeling based on the situation
 
-**Let personality shape the reasoning.** Personality (traits + dimensions) should color the situation read, which goals feel salient, how the plan is approached, and the emotional tone. A high-curiosity NPC notices different things than a low-curiosity one; an "idiotic" NPC reasons differently than a "meticulous" one. Make the personality legible in the assessment, goals, plan, and emotional state rather than producing generic, trait-agnostic reasoning.
+**Let personality shape the reasoning.** Personality (traits + dimensions) should color the situation assessment, which goals feel salient, how the plan is approached, and the emotional tone. A high-curiosity NPC notices different things than a low-curiosity one; an "idiotic" NPC reasons differently than a "meticulous" one. Make the personality legible in the assessment, goals, plan, and emotional state rather than producing generic, trait-agnostic reasoning.
 
 **Handle interaction lifecycle events:**
 - **"Interaction finished / canceled: X"** → Clear any goals/plans related to X, acknowledge completion, assess what to do next
